+++ DiscordBotNet/LegendaryBot/Quests/DirectionHelping.cs
@@ -96,7 +96,7 @@ public class DirectionHelping : Quest
 
         var userData = await databaseContext.UserData
             .IncludeTeamWithAllEquipments()
-            .FindOrCreateAsync(context.User.Id);
+            .FindOrCreateAsync((long)context.User.Id);
         var userTeam = userData.EquippedPlayerTeam;
         await userTeam.LoadAsync(context.User);
         var battle = new BattleSimulator(userTeam,blastTeam);
