+++ DiscordBotNet/LegendaryBot/command/Info.cs
@@ -16,7 +16,7 @@ public class Info : BaseCommandClass
             
         if(author is null)author = ctx.User;
         
-        var userData = await DatabaseContext.UserData.FindOrCreateAsync(author.Id);
+        var userData = await DatabaseContext.UserData.FindOrCreateAsync((long)author.Id);
   
         DiscordEmbedBuilder embedBuilder = new DiscordEmbedBuilder()
             .WithTitle("Info")
